requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
requests